NL-008 modificacion logica servidor para manejo de wars remotos

+++ servidor.js
@@ -5,6 +5,7 @@ const fs = require("fs");
 const unrar = require("node-unrar-js");
 const rimraf = require("rimraf");
 const { exec } = require("child_process");
+const multer = require("multer");
 
 const app = express();
 app.use(cors());
@@ -43,30 +44,56 @@ app.use(bodyParser.json());
   });
 }); */
 
-app.post("/archivo-war", async (req, res) => {
-  const { filePath, newContent, contentLine } = req.body;
-  const rarFilePath = filePath.replace(".war", ".rar");
+const upload = multer();
+app.post("/archivo-war", upload.single("warData"), async (req, res) => {
+  const { newContent, contentLine } = req.body;
+  const warData = req.file.buffer;
+
+  const warFilePath = "temp.war";
+  fs.writeFileSync(warFilePath, Buffer.from(warData, "base64"));
+
+  const rarFilePath = warFilePath.replace(".war", ".rar");
+  fs.renameSync(warFilePath, rarFilePath);
 
   try {
-    fs.renameSync(filePath, rarFilePath);
-
     const result = await modificarArchivoEnRAR(
       rarFilePath,
       newContent,
       contentLine
     );
     console.log(result);
-    res.status(200).json({
-      message: "good",
-      success: true,
+
+    const warFilePath = rarFilePath.replace(".rar", ".war");
+    fs.renameSync(rarFilePath, warFilePath);
+
+    const modifiedWarData = fs.readFileSync(warFilePath);
+
+    res.status(result.status).json({
+      message: result.message,
+      success: result.success,
+      modifiedWarData:
+        result.status == 200 ? modifiedWarData.toString("base64") : null,
     });
   } catch (error) {
     console.log(error);
-    fs.renameSync(rarFilePath, rarFilePath.replace(".rar", ".war"));
+    if (fs.existsSync(rarFilePath)) {
+      fs.renameSync(rarFilePath, rarFilePath.replace(".rar", ".war"));
+    }
+
     res.status(500).json({
       message: "Error interno del servidor",
       success: false,
     });
+  } finally {
+    if (fs.existsSync(warFilePath)) {
+      fs.unlinkSync(warFilePath);
+    }
+    if (fs.existsSync(rarFilePath)) {
+      fs.unlinkSync(rarFilePath);
+    }
+    if (fs.existsSync(rarFilePath)) {
+      fs.unlinkSync(rarFilePath);
+    }
   }
 });
 
@@ -90,7 +117,13 @@ async function modificarArchivoEnRAR(rarFilePath, newContent, contentLine) {
       let nuevaCadena = cadena.substring(0, indiceUltimaBarra + 1);
       fs.mkdirSync("temp/" + nuevaCadena, { recursive: true });
       fs.writeFileSync("temp/" + file.fileHeader.name, intento, "utf8");
-    } catch (error) {}
+    } catch (error) {
+      return {
+        status: 500,
+        message: "error al leer el archivo",
+        success: false,
+      };
+    }
   });
 
   Promise.all(promises);
@@ -104,26 +137,52 @@ async function modificarArchivoEnRAR(rarFilePath, newContent, contentLine) {
         if (err) {
           rimraf.sync(directorio);
           fs.renameSync(rarFilePath, rarFilePath.replace(".rar", ".war"));
-          reject(`Error al leer la carpeta: ${error.message}`);
+          reject({
+            status: 500,
+            message: `Error al leer la carpeta: ${error.message}`,
+            success: false,
+          });
         }
         // const comando = '"C:\\Program Files\\WinRAR\\WinRAR.exe" a -r -ep1 prueba.rar temp/*'; -> comando para no tener winrar en el server
 
-        const comando = "winrar a -r -ep1 prueba.rar temp/*";
+        const comando = "winrar a -r -ep1 temp2.rar temp/*";
 
         exec(comando, (error, stdout, stderr) => {
           if (error) {
             rimraf.sync(directorio);
+            console.log(
+              "ERRORHP___________________________________________________________________________________________"
+            );
+            console.log(error);
             fs.renameSync(rarFilePath, rarFilePath.replace(".rar", ".war"));
-            reject(`Error al comprimir la carpeta: ${error.message}`);
+            reject({
+              status: 500,
+              message: `Error al comprimir la carpeta: ${error.message}`,
+              success: false,
+            });
           }
           if (stderr) {
             rimraf.sync(directorio);
+            console.log(
+              "ERRORHP2___________________________________________________________________________________________"
+            );
+            console.log(stderr);
             fs.renameSync(rarFilePath, rarFilePath.replace(".rar", ".war"));
-            reject(`Error de WinRAR: ${stderr}`);
+            reject({
+              status: 500,
+              message: `Error de procesamiento: ${stderr}`,
+              success: false,
+            });
           }
+
+          fs.renameSync("temp2.rar", "temp.rar");
+
           rimraf.sync(directorio);
-          fs.renameSync(rarFilePath, rarFilePath.replace(".rar", ".war"));
-          resolve("El archivo dentro del RAR ha sido modificado exitosamente.");
+          resolve({
+            status: 200,
+            message: "Datos actualizados correctamente",
+            success: true,
+          });
         });
       });
     });
